Seperated files, also sending shipping data to backend
